feat: BSA change notification, type-ordered insert, and researched field names
Model
- BSA_Entry, BSA_Collision and BSA_Expiration implement INotifyPropertyChanged
  so editors can bind to them directly instead of rebuilding their views on
  every edit. Only the properties that back a displayed value use backing
  fields. The events are NonSerialized so BinaryFormatter copies and clipboard
  use keep working.
- AddEntry(int, BSA_Entry) assigns the requested ID and rejects duplicates.
  It previously ignored the ID on the add path, so entries added at a specific
  ID always landed at 0, and its replace branch destroyed an entry with no undo
  record. Duplicate IDs cannot be written: the deserializer sizes the main entry
  pointer table by unique ID but writes bodies by list position.
- AddIBsaType inserts a subtype at the end of its own TypeID block. The format
  writes one header per type, so cross-type order cannot round-trip. Keeping the
  list in TypeID order makes it match what is saved.

Researched names, from tracing how the game reads these fields
- Collision and Expiration serialize as CollisionEffect and CollisionSound.
- BSA_Expiration I_00 and I_04 are an ACB type and cue ID.
- Type 2, 12, 13 and 14 serialize under names describing what they do.
- Type 4 I_00 and Type 8 I_00 are enums rather than opaque values.
- Type 13 I_12 and I_20, and Type 14 F_04 and F_52, are read at the width the
  game uses rather than the width previously assumed.

+++ Xv2CoreLib/BSA_XV1/Deserializer.cs
@@ -145,7 +145,7 @@ namespace Xv2CoreLib.BSA_XV1
 
                 for (int i = 0; i < unk2.Count(); i++)
                 {
-                    bytes.AddRange(BitConverter.GetBytes(unk2[i].I_00));
+                    bytes.AddRange(BitConverter.GetBytes((ushort)unk2[i].I_00));
                     bytes.AddRange(BitConverter.GetBytes(unk2[i].I_02));
                     bytes.AddRange(BitConverter.GetBytes(unk2[i].I_04));
                     bytes.AddRange(BitConverter.GetBytes(unk2[i].I_06));
